refactor: split App.tsx into focused components and hooks (#481)
- Add useChatSession hook for chat state, history, and handlers
- Add useTourState hook for Joyride tour state and steps
- Add useAppOrchestration hook for top-level state wiring
- Add AppShell component for main layout structure
- Reduce App.tsx from 1,612 to ~278 lines
- Set coverage thresholds to 85% (lines/functions), 78% (branches)
- Add tests for help-cache, nanoid non-secure, password, useTourState,
  useChatSession, useAppOrchestration

Made-with: Cursor
